refactor(validation): promote core/validation/ → validation/ (top-level)
ValidationRule and ValidationResult are public API.
Update 4 import references. Remove validation from core/__init__.py.
