+++ src/bootstack/validation/validation_result.py
